Fixed some minor issues

+++ RestLS/RestLS/Data/Dtos/Appoitments/AppointmentDto.cs
@@ -1,5 +1,5 @@
 ﻿namespace RestLS.Data.Dtos.Appoitments;
 
-public record AppointmentDto(int Id, DateTime Time, double Price, int DocId);
-public record CreateAppointmentDto(double Price, string Time);
-public record UpdateAppointmentDto(string Time, double Price);
+public record AppointmentDto(int Id, DateTime Time, decimal Price, int DocId);
+public record CreateAppointmentDto(decimal Price, string Time);
+public record UpdateAppointmentDto(string Time, decimal Price);
