+++ RestLS/RestLS/Data/Dtos/GroupSessions/GroupSessionsDto.cs
@@ -2,5 +2,5 @@
 
 
 public record GroupSessionDto(int Id, string Name, int DocId);
-public record CreateGroupSessionDto(string Name, string Description,double Price,int Spaces, string Time);
-public record UpdateGroupSessionDto(string Name, string Description,double Price,int Spaces, string Time);
+public record CreateGroupSessionDto(string Name, string Description,decimal Price,int Spaces, string Time);
+public record UpdateGroupSessionDto(string Name, string Description,decimal Price,int Spaces, string Time);
